refactor: databaseInitializer.tsを責務ごとに分割
1432行のモノリシックファイルを以下に分割:
- databaseInitializer.ts (131行) - オーケストレーション
- schema/migrationSql.ts - SQL DDL定数
- schema/migrationRunner.ts - マイグレーション実行ロジック
- databaseHealth.ts - ヘルスチェック・最適化
- databaseUtils.ts - ヘルパー関数

全てのimport元を新しいパスに直接更新。

Closes #671
