+++ electron-src/lib/prisma/databaseHealth.ts
@@ -1,0 +1,40 @@
+import { createSharedPrismaClient } from "./databaseInitializer"
+
+// データベースの健全性チェック
+export const checkDatabaseHealth = async (): Promise<boolean> => {
+  const prisma = createSharedPrismaClient()
+
+  try {
+    // 簡単なクエリで接続確認
+    await prisma.$queryRaw`SELECT 1`
+    return true
+  } catch (error) {
+    console.error("Database health check failed:", error)
+    return false
+  } finally {
+    await prisma.$disconnect()
+  }
+}
+
+// 共有ドライブ用のSQLite最適化設定
+export const optimizeDatabaseForSharedDrive = async (): Promise<void> => {
+  const prisma = createSharedPrismaClient()
+
+  try {
+    // WALモードを有効にして同時読み取りを改善
+    await prisma.$queryRaw`PRAGMA journal_mode = WAL`
+
+    // 読み取り専用トランザクションのタイムアウトを短縮
+    await prisma.$queryRaw`PRAGMA busy_timeout = 30000`
+
+    // 同期モードを調整（共有ドライブでのパフォーマンス向上）
+    await prisma.$queryRaw`PRAGMA synchronous = NORMAL`
+
+    // キャッシュサイズを増加
+    await prisma.$queryRaw`PRAGMA cache_size = -64000`
+  } catch (error) {
+    console.error("Failed to optimize database:", error)
+  } finally {
+    await prisma.$disconnect()
+  }
+}
